+++ types/index.ts
@@ -2,5 +2,6 @@ export * from "./api";
 export * from "./auth";
 export * from "./backup";
 export * from "./cache";
+export * from "./scrobble";
 export * from "./theme";